// src/infrastructure/api/generated/model/adoptionSummaryResponseDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Pet Rescue API
 * RESTful API for Pet Rescue and Adoption Platform
 * OpenAPI spec version: 1.0.0
 */
import type { AdoptionSummaryResponseDtoStatus } from './adoptionSummaryResponseDtoStatus';

/**
 * Adoption application summary for list views
 */
export interface AdoptionSummaryResponseDto {
  applicationId?: string;
  adoptionCode?: string;
  petName?: string;
  petPrimaryImageUrl?: string;
  applicantUsername?: string;
  status?: AdoptionSummaryResponseDtoStatus;
  experience?: string;
  liveCondition?: string;
  createdAt?: string;
}